+++ test/main.jl
@@ -198,6 +198,19 @@ end
     @test true
 end
 
+@testitem "install/remove python (build=**cpython**)" begin
+    @testset "$version" for version in ["3.12.*", "3.13.*"]
+        include("setup.jl")
+        CondaPkg.add("python", version = version, build = "**cpython**")
+        CondaPkg.resolve()
+        @test occursin("python", status())
+        @test occursin("($version, build=**cpython**)", status())
+        CondaPkg.rm("python")
+        CondaPkg.resolve()
+        @test !occursin("python", status())
+    end
+end
+
 @testitem "install non-existent package" begin
     include("setup.jl")